Fix component renaming and capitalize week
