--- src/components/theme/Theme.css
@@ -1,9 +1,0 @@
-/* Header styles */
-.theme {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-h3 {
-  margin: 10px;
-}
