+++ src/components/title/Title.css
@@ -1,0 +1,23 @@
+/* Header styles within Reading and Response */
+.header {
+  display: flex;
+  justify-content: space-between;
+}
+
+/* Individual Reading and Response text styles */
+.title {
+  text-align: left;
+  margin: 0;
+}
+
+.verse {
+  text-align: right;
+  margin: 0;
+}
+
+
+.summary {
+  text-align: center;
+  margin: 0;
+  font-style: italic;
+}
